requirements.txt
termcolor
tqdm
pyyaml
easydict
torch
torchvision
pennylane
qiskit
pennylane-qiskit
tensorflow
tensorflow-quantum